blog styling
created at timestamp

+++ portfolio-app/src/components/Posts.js
@@ -8,25 +8,40 @@ import { NColumnContent } from './Containers';
 import ColorButton from "./ColorButton";
 
 const PostsContainer = styled.div`
-  width: 100%;
   height: 2000px;
   align-items: center;
   display: flex;
   flex-direction: column;
 `
 const Post = styled.div`
+  display: flex;
+  flex-direction: column;
   color: black;
-  border-radius: 25px;
+  border-radius: 3px;
   padding: 50px;
-  border: 5px solid red;
+  border: 2px solid black;
+`
+
+const CreatedAt = styled.div`
+  padding: 0px;
+  margin: 0px;
+  opacity:0.3;
 `
 
 const PostContainer = styled.div`
   border-radius: 25px;
-  padding: 50px;
+  padding: 16px;
 `
 
-function Posts (props) {
+const formatDate = (time_stamp) => {
+  const date = new Date(time_stamp);
+  const dateTimeFormat = new Intl.DateTimeFormat('en', { year: 'numeric', month: 'short', day: '2-digit' }) 
+  const [{ value: month },,{ value: day },,{ value: year }] = dateTimeFormat .formatToParts(date)
+
+  return `${day}-${month}-${year }`
+}
+
+const Posts = (props) => {
   const arr = [1, 3, 5, 2, 7, 9, 2, 2, 1]
   const dispatch = useDispatch();
   const postsError = useSelector(state => state.PostsReducer.fetchPostsError)
@@ -51,16 +66,22 @@ function Posts (props) {
               <Post >
                 {isAdmin ? 
                 <ColorButton  
+                margins="0px"
                 color='#FF0000'
                 text='Delete'
                 onClick={() => {
+                  if (window.confirm(`Do you really want to delete ${post.title} ?`) == true) { 
                   dispatch(deletePost(post.id))
+                  }
                 }}
                 />: undefined}
-              
-                <Title>
+                <Title size="50px" margins="margin: 0px;">
                   {post.title}
                 </Title>
+                <CreatedAt>
+                Created on: {formatDate(post.time_stamp)}
+                </CreatedAt>
+               
                 <div dangerouslySetInnerHTML={{__html: post.content}}/>
                 
               </Post>
